test: Add Get Balance test
